demo/src/tracing/requirements.txt
opencensus
opencensus-ext-ocagent
opencensus-ext-flask
opencensus-ext-requests
flask